♻️(backend) refactor document parsers
I refactor document parsing by introducing
AlbertParser and BaseParser
