+++ src/backend/chat/agent_rag/document_rag_backends/albert_rag_backend.py
@@ -13,7 +13,7 @@ import requests
 
 from chat.agent_rag.albert_api_constants import Searches
 from chat.agent_rag.constants import RAGWebResult, RAGWebResults, RAGWebUsage
-from chat.agent_rag.document_converter.markitdown import DocumentConverter
+from chat.agent_rag.document_converter.parser import AlbertParser
 from chat.agent_rag.document_rag_backends.base_rag_backend import BaseRagBackend
 
 logger = logging.getLogger(__name__)
@@ -46,10 +46,9 @@ class AlbertRagBackend(BaseRagBackend):  # pylint: disable=too-many-instance-att
         }
         self._collections_endpoint = urljoin(self._base_url, "/v1/collections")
         self._documents_endpoint = urljoin(self._base_url, "/v1/documents")
-        self._pdf_parser_endpoint = urljoin(self._base_url, "/v1/parse-beta")
         self._search_endpoint = urljoin(self._base_url, "/v1/search")
-
         self._default_collection_description = "Temporary collection for RAG document search"
+        self.parser = AlbertParser()
 
     def create_collection(self, name: str, description: Optional[str] = None) -> str:
         """
@@ -113,58 +112,6 @@ class AlbertRagBackend(BaseRagBackend):  # pylint: disable=too-many-instance-att
                 timeout=settings.ALBERT_API_TIMEOUT,
             )
             response.raise_for_status()
-
-    def parse_pdf_document(self, name: str, content_type: str, content: BytesIO) -> str:
-        """
-        Parse the PDF document content and return the text content.
-        This method should handle the logic to convert the PDF into
-        a format suitable for the Albert API.
-        """
-        response = requests.post(
-            self._pdf_parser_endpoint,
-            headers=self._headers,
-            files={
-                "file": (
-                    name,
-                    content,
-                    content_type,
-                ),  # Use the name as the filename in the request
-                "output_format": (None, "markdown"),  # Specify the output format as Markdown,
-            },
-            timeout=settings.ALBERT_API_PARSE_TIMEOUT,
-        )
-        response.raise_for_status()
-
-        return "\n\n".join(
-            document_page["content"] for document_page in response.json().get("data", [])
-        )
-
-    def parse_document(self, name: str, content_type: str, content: BytesIO):
-        """
-        Parse the document and prepare it for the search operation.
-        This method should handle the logic to convert the document
-        into a format suitable for the Albert API.
-
-        Args:
-            name (str): The name of the document.
-            content_type (str): The MIME type of the document (e.g., "application/pdf").
-            content (BytesIO): The content of the document as a BytesIO stream.
-
-        Returns:
-            str: The document content in Markdown format.
-        """
-        # Implement the parsing logic here
-        if content_type == "application/pdf":
-            # Handle PDF parsing
-            markdown_content = self.parse_pdf_document(
-                name=name, content_type=content_type, content=content
-            )
-        else:
-            markdown_content = DocumentConverter().convert_raw(
-                name=name, content_type=content_type, content=content
-            )
-
-        return markdown_content
 
     def store_document(self, name: str, content: str, **kwargs) -> None:
         """
